// superset-frontend/src/dashboard/types.ts
export type Chart = ChartState & {
  form_data: {
    viz_type: string;
    datasource: string;
    color_scheme: string;
    slice_id: number;
  };
};
// ...
export type DashboardLayout = { [key: string]: LayoutItem };
// ...
export type ChartsState = { [key: string]: Chart };
// ...
export type ComponentType = (typeof componentTypes)[ComponentTypesKeys];

export type LayoutItemMeta = {
  chartId: number;
  defaultText?: string;
  height: number;
  placeholder?: string;
  sliceName?: string;
  sliceNameOverride?: string;
  text?: string;
  uuid: string;
  width: number;
};

/** State of dashboardLayout item in redux */
export type LayoutItem = {
  children: string[];
  parents?: string[];
  type: ComponentType;
  id: string;
  meta: LayoutItemMeta;
};